+++ cli/src/projectfs.rs
@@ -3,7 +3,7 @@ use std::{fs, io};
 
 use path_abs::{PathAbs, PathDir};
 
-use crate::materials::toml;
+use crate::transpiler::toml;
 
 pub struct Project {
     pub name: String,
